+++ modules/apps/archived/export-import-resources-importer/src/main/java/com/liferay/exportimport/resources/importer/internal/messaging/ResourcesImporterHotDeployMessageListener.java
@@ -16,9 +16,6 @@ import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.messaging.Destination;
-import com.liferay.portal.kernel.messaging.DestinationConfiguration;
-import com.liferay.portal.kernel.messaging.DestinationFactory;
 import com.liferay.portal.kernel.messaging.HotDeployMessageListener;
 import com.liferay.portal.kernel.messaging.Message;
 import com.liferay.portal.kernel.messaging.MessageBus;
@@ -29,15 +26,11 @@ import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
 import com.liferay.portal.kernel.service.CompanyLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
-import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 import com.liferay.portal.kernel.util.Validator;
-
-import java.util.Dictionary;
 
 import javax.servlet.ServletContext;
 
 import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Deactivate;
@@ -56,8 +49,6 @@ public class ResourcesImporterHotDeployMessageListener
 
 	@Activate
 	protected void activate(BundleContext bundleContext) {
-		_bundleContext = bundleContext;
-
 		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
 			bundleContext, ServletContext.class, null,
 			(serviceReference, emitter) -> {
@@ -74,31 +65,11 @@ public class ResourcesImporterHotDeployMessageListener
 					bundleContext.ungetService(serviceReference);
 				}
 			});
-
-		DestinationConfiguration destinationConfiguration =
-			new DestinationConfiguration(
-				DestinationConfiguration.DESTINATION_TYPE_SERIAL,
-				ResourcesImporterDestinationNames.RESOURCES_IMPORTER);
-
-		_destination = _destinationFactory.createDestination(
-			destinationConfiguration);
-
-		Dictionary<String, Object> dictionary =
-			HashMapDictionaryBuilder.<String, Object>put(
-				"destination.name", _destination.getName()
-			).build();
-
-		_serviceRegistration = _bundleContext.registerService(
-			Destination.class, _destination, dictionary);
 	}
 
 	@Deactivate
 	protected void deactivate() {
-		_serviceRegistration.unregister();
-
 		_serviceTrackerMap.close();
-
-		_bundleContext = null;
 	}
 
 	@Override
@@ -239,15 +210,8 @@ public class ResourcesImporterHotDeployMessageListener
 	private static final Log _log = LogFactoryUtil.getLog(
 		ResourcesImporterHotDeployMessageListener.class);
 
-	private BundleContext _bundleContext;
-
 	@Reference
 	private CompanyLocalService _companyLocalService;
-
-	private Destination _destination;
-
-	@Reference
-	private DestinationFactory _destinationFactory;
 
 	@Reference
 	private ImporterFactory _importerFactory;
@@ -260,7 +224,6 @@ public class ResourcesImporterHotDeployMessageListener
 	)
 	private Release _release;
 
-	private ServiceRegistration<Destination> _serviceRegistration;
 	private ServiceTrackerMap<String, ServletContext> _serviceTrackerMap;
 
 }
