--- modules/apps/archived/export-import-resources-importer/src/main/java/com/liferay/exportimport/resources/importer/internal/messaging/ResourcesImporterMessagingConfigurator.java
@@ -1,60 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.exportimport.resources.importer.internal.messaging;
-
-import com.liferay.portal.kernel.deploy.hot.HotDeployListener;
-import com.liferay.portal.kernel.deploy.hot.HotDeployUtil;
-import com.liferay.portal.kernel.messaging.Destination;
-import com.liferay.portal.kernel.messaging.DestinationConfiguration;
-import com.liferay.portal.kernel.messaging.DestinationFactory;
-import com.liferay.portal.kernel.util.MapUtil;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.ServiceRegistration;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Shuyang Zhou
- */
-@Component(service = {})
-public class ResourcesImporterMessagingConfigurator {
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		DestinationConfiguration destinationConfiguration =
-			new DestinationConfiguration(
-				DestinationConfiguration.DESTINATION_TYPE_SERIAL,
-				DestinationNames.HOT_DEPLOY);
-
-		Destination destination = _destinationFactory.createDestination(
-			destinationConfiguration);
-
-		_serviceRegistration = bundleContext.registerService(
-			Destination.class, destination,
-			MapUtil.singletonDictionary(
-				"destination.name", destination.getName()));
-
-		HotDeployUtil.registerListener(_hotDeployListener);
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		HotDeployUtil.unregisterListener(_hotDeployListener);
-
-		_serviceRegistration.unregister();
-	}
-
-	@Reference
-	private DestinationFactory _destinationFactory;
-
-	private final HotDeployListener _hotDeployListener =
-		new MessagingHotDeployListener();
-	private ServiceRegistration<Destination> _serviceRegistration;
-
-}
